+++ app/helpers/modal_helper.rb
@@ -2,7 +2,12 @@
 
 module ModalHelper
   MODAL_ID     = "modal"
-  LINK_OPTIONS = { data: { turbo_frame: MODAL_ID, action: "modal#onLinkClick", modal_type: "" } }.freeze
+  LINK_OPTIONS = { data: {
+    turbo_frame: MODAL_ID,
+    controller:  "modal-link",
+    action:      "modal-link#onClick",
+    modal_type:  "",
+  } }.freeze
 
   def modal_tag(modal_id = MODAL_ID)
     turbo_frame_tag modal_id, data: { controller: "modal", modal_target: "turboFrame", action: <<~ACTIONS }
@@ -11,22 +16,16 @@ module ModalHelper
     ACTIONS
   end
 
+  # @see ActionView::Helpers::UrlHelper#link_to for argument details
   def modal_link_to(name = nil, options = nil, html_options = nil, &block)
     if block
-      link_to name, LINK_OPTIONS.deep_merge(options || {}), html_options, &block
+      link_to name, LINK_OPTIONS.deep_merge(options || {}), &block
     else
-      link_to name, options, LINK_OPTIONS.deep_merge(html_options || {}), &block
+      link_to name, options, LINK_OPTIONS.deep_merge(html_options || {})
     end
   end
 
-  def modal_link(text = nil, path = nil, options = {}, &block)
-    options = LINK_OPTIONS.deep_merge(options)
-    if block
-      link_to text || path, options, &block
-    else
-      link_to text, path, options
-    end
-  end
+  alias modal_link modal_link_to
 
   def modal_content(options = {}, &block)
     modal_id = options.fetch(:modal_id, MODAL_ID)