+++ app/javascript/turbo_modal/index.js
@@ -1,4 +1,16 @@
 import ModalController, { openModal, closeModal } from "./modal_controller";
+import ModalLinkController from "./modal_link_controller";
 import ScrimController, { showScrim, hideScrim } from "./scrim_controller";
 
-export { ModalController, openModal, closeModal, ScrimController, showScrim, hideScrim };
+const Definitions = [
+  { identifier: "modal", controllerConstructor: ModalController },
+  { identifier: "modal-link", controllerConstructor: ModalLinkController },
+  { identifier: "scrim", controllerConstructor: ScrimController }
+];
+
+export {
+  Definitions as default,
+  ModalController, openModal, closeModal,
+  ScrimController, showScrim, hideScrim,
+  ModalLinkController
+};